+++ src/restfw/authorization.py
@@ -10,8 +10,15 @@ from pyramid.security import ACLAllowed, ACLDenied, ALL_PERMISSIONS, Allow, Deny
 from zope.interface import implementer
 
 
+ALL_GET_REQUESTS = 'get'
+ALL_POST_REQUESTS = 'post'
+ALL_PUT_REQUESTS = 'put'
+ALL_PATCH_REQUESTS = 'patch'
+ALL_DELETE_REQUESTS = 'delete'
+
+
 @implementer(IAuthorizationPolicy)
-class RestACLAuthorizationPolicy(object):
+class RestACLAuthorizationPolicy:
     """ An :term:`authorization policy` which consults an :term:`ACL`
     object attached to a :term:`context` to determine authorization
     information about a :term:`principal` or multiple principals.
@@ -62,14 +69,7 @@ class RestACLAuthorizationPolicy(object):
     :class:`pyramid.interfaces.IAuthorizationPolicy` interface.
     """
 
-    @staticmethod
-    def _get_context_permission(context, permission):
-        """Returns context-aware version of given permission."""
-        if context:
-            method_options = getattr(context, 'options_for_%s' % permission, None)
-            if method_options is not None and method_options.permission is not None:
-                permission = method_options.permission
-        return permission
+    BASE_PERMISSIONS = {'get', 'post', 'put', 'patch', 'delete'}
 
     @staticmethod
     def _match_permission(permission, ace_permissions):
@@ -85,15 +85,24 @@ class RestACLAuthorizationPolicy(object):
                 return True
         return False
 
-    def permits(self, context, principals, permission):
+    @staticmethod
+    def _match_base_permission(permission, ace_permissions):
+        """Match base permission with list of permissions from ACE (Access Control Entries).
+        """
+        if ace_permissions is ALL_PERMISSIONS:
+            return True
+        return any(permission == ace_permission for ace_permission in ace_permissions)
+
+    def permits(self, context, principals, permission: str):
         """ Return an instance of
         :class:`pyramid.security.ACLAllowed` instance if the policy
         permits access, return an instance of
         :class:`pyramid.security.ACLDenied` if not."""
-
         acl = '<No ACL found on any object in resource lineage>'
-        context_permission = self._get_context_permission(context, permission)
-        orig_permission = permission if permission != context_permission else None
+        base_permission, _, context_permission = permission.partition('.')
+        if base_permission and base_permission not in self.BASE_PERMISSIONS:
+            context_permission = permission
+            base_permission = ''
         permission = context_permission
 
         for location in lineage(context):
@@ -111,7 +120,7 @@ class RestACLAuthorizationPolicy(object):
                     if not is_nonstr_iter(ace_permissions):
                         ace_permissions = [ace_permissions]
                     if (self._match_permission(permission, ace_permissions) or
-                            (orig_permission and self._match_permission(orig_permission, ace_permissions))):
+                            (base_permission and self._match_base_permission(base_permission, ace_permissions))):
                         if ace_action == Allow:
                             return ACLAllowed(ace, acl, permission,
                                               principals, location)
@@ -129,13 +138,17 @@ class RestACLAuthorizationPolicy(object):
             context
         )
 
-    def principals_allowed_by_permission(self, context, permission):
+    def principals_allowed_by_permission(self, context, permission: str):
         """ Return the set of principals explicitly granted the
         permission named ``permission`` according to the ACL directly
         attached to the ``context`` as well as inherited ACLs based on
         the :term:`lineage`."""
         allowed = set()
-        permission = self._get_context_permission(context, permission)
+        base_permission, _, context_permission = permission.partition('.')
+        if base_permission and base_permission not in self.BASE_PERMISSIONS:
+            context_permission = permission
+            base_permission = ''
+        permission = context_permission
 
         for location in reversed(list(lineage(context))):
             # NB: we're walking *up* the object graph from the root
@@ -153,7 +166,10 @@ class RestACLAuthorizationPolicy(object):
             for ace_action, ace_principal, ace_permissions in acl:
                 if not is_nonstr_iter(ace_permissions):
                     ace_permissions = [ace_permissions]
-                is_match = self._match_permission(permission, ace_permissions)
+                is_match = (
+                        self._match_permission(permission, ace_permissions)
+                        or (base_permission and self._match_base_permission(base_permission, ace_permissions))
+                )
                 if (ace_action == Allow) and is_match:
                     if ace_principal not in denied_here:
                         allowed_here.add(ace_principal)